crypto: qat - move fw config related structures
commit 98a4f29fba0f ("crypto: qat - move fw config related structures") upstream

Relocate the structures adf_fw_objs and adf_fw_config from the file
adf_4xxx_hw_data.c to the newly created adf_fw_config.h.

These structures will be used by new device drivers.

This does not introduce any functional change.

Intel-SIG: commit 98a4f29fba0f ("crypto: qat - move fw config related structures)"
Backport to support Intel QAT in-tree driver

[ Aichun Shi: amend commit log ]
